add vitest for browser testing
