+++ package.json
@@ -76,6 +76,7 @@
     "@types/jasmine": "4.0.3",
     "@typescript-eslint/eslint-plugin": "5.33.0",
     "@typescript-eslint/parser": "5.33.0",
+    "@vitest/browser": "^3.0.5",
     "babel-loader": "8.2.5",
     "concurrently": "7.3.0",
     "eslint": "8.21.0",
@@ -93,12 +94,14 @@
     "karma-typescript": "5.5.3",
     "karma-typescript-es6-transform": "5.5.3",
     "loki": "0.34.0",
+    "playwright": "^1.50.1",
     "prettier": "2.7.1",
     "rimraf": "3.0.2",
     "rollup-plugin-visualizer": "^5.14.0",
     "ts-loader": "9.3.1",
     "typescript": "4.7.4",
     "vite": "^6.1.0",
+    "vitest": "^3.0.5",
     "webpack": "5.76.0",
     "webpack-bundle-analyzer": "4.5.0",
     "webpack-cli": "4.10.0",
