+++ src/utils/customElements.ts
@@ -1,12 +1,11 @@
-import { DataAttribute, DisableableFeature, MetadataAttribute } from './dataAttributes';
+import {
+  DataAttribute,
+  DisableableFeature,
+  getHighlightTypeForElement,
+  HighlightType,
+  MetadataAttribute,
+} from './dataAttributes';
 import { ConfigurationManager } from '../lib/ConfigurationManager';
-
-export enum HighlightType {
-  None = '',
-  Element = 'element',
-  ContentComponent = 'content-component',
-  ContentItem = 'content-item',
-}
 
 const DisabledHighlightFeatureSelector = `[${MetadataAttribute.DisableFeatures}*="${DisableableFeature.Highlight}"]`;
 const ElementSelector = `*[${DataAttribute.ElementCodename}]:not(${DisabledHighlightFeatureSelector})`;
@@ -77,36 +76,4 @@ export function shouldElementHaveAddButton(element: HTMLElement | null): boolean
   const isInsideWebSpotlight = configurationManager.isInsideWebSpotlightPreviewIFrame;
   // add button should only be visible inside Web Spotlight
   return Boolean(isInsideWebSpotlight && element && element.hasAttribute(MetadataAttribute.AddButton));
-}
-
-/**
- * Get HighlightType based on the HTML element data-attributes.
- *
- * @param {HTMLElement | null} element
- * @returns {HighlightType}
- */
-export function getHighlightTypeForElement(element: HTMLElement | null): HighlightType {
-  if (!element || isFeatureDisabledForElement(element, DisableableFeature.Highlight)) {
-    return HighlightType.None;
-  }
-
-  // treat node as element if it has element codename attribute
-  if (element.hasAttribute(DataAttribute.ElementCodename)) {
-    return HighlightType.Element;
-  }
-  // else treat node as content component if it has component id attribute
-  else if (element.hasAttribute(DataAttribute.ComponentId)) {
-    return HighlightType.ContentComponent;
-  }
-  // else treat node as content item if it has item id attribute
-  else if (element.hasAttribute(DataAttribute.ItemId)) {
-    return HighlightType.ContentItem;
-  }
-
-  return HighlightType.None;
-}
-
-function isFeatureDisabledForElement(element: HTMLElement, feature: DisableableFeature): boolean {
-  const attribute = element.getAttribute(MetadataAttribute.DisableFeatures);
-  return attribute?.toLowerCase().includes(feature) ?? false;
 }
