+++ test-browser/utils/assert.spec.ts
@@ -1,9 +1,10 @@
+import { describe, it, expect, vi } from 'vitest';
 import { assert } from '../../src/utils/assert';
 import { Logger } from '../../src/lib/Logger';
 
 describe('assert.ts', () => {
   it('should throw error if condition is falsy', () => {
-    spyOn(Logger, 'error');
+    vi.spyOn(Logger, 'error');
 
     expect(() => {
       assert(false, 'error-message');
