+++ vitest.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    alias: { '@/': new URL('./src/', import.meta.url).pathname },
+    root: './test-browser',
+    browser: {
+      enabled: true,
+      provider: 'playwright',
+      screenshotFailures: false,
+      instances: [{ browser: 'chromium' }],
+    },
+  },
+});
